1-6

+++ additional_tasks/find_sum_numbers/src/utils.cpp
@@ -3,8 +3,8 @@
 #include <algorithm>
 
 std::pair<int, int> GetNumbersOfSum(const std::vector<int>& numbers, int sum) {
-  int i = 0;
-  int j = numbers.size() - 1;
+  size_t i = 0;
+  size_t j = numbers.size() - 1;
   while (i < j) {
     if (numbers[i] + numbers[j] == sum) {
       return {numbers[i], numbers[j]};
